+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- Add `connection_type` to the device context of events, which reports `wifi`, `ethernet`, `cellular`, or `none`. On iOS, cellular connections also report the network technology, for example `cellular_4g` or `cellular_5g` (#9097)
+- Add `connection_type` to the device context of events, which reports `wifi`, `ethernet`, `cellular`, or `none`. On iOS, cellular connections additionally report the network technology in `connection_effective_type`, for example `4g` or `5g` (#9097)
 
 ### Fixes